fixed quest error

+++ help_threepwood_quest_chain.gd
@@ -26,6 +26,8 @@ func entered_pickup_zone(body: Node3D) -> void:
 		quest_marker.modulate = Color.GRAY
 		state_chart.send_event("pickup_quest")
 		pickup_zone.body_entered.disconnect(entered_pickup_zone)
+		state_chart.set_expression_property("target_killed", false)
+		state_chart.set_expression_property("item_picked_up", false)
 
 func kill_target_entered() -> void:
 	print("kill target dummy")
@@ -39,11 +41,12 @@ func kill_target_entered() -> void:
 func pickup_coin_entered() -> void:
 	print("pick up coin")
 	pickup_coin.set_visibility.call_deferred(true)
-	pickup_coin.picked_up.connect(func():
-		QuestManager.update_description("You found the coin! Now go and kill the Target Dummy.")
-		print("picked up coin")
-		state_chart.set_expression_property("item_picked_up", true)
-		state_chart.send_event("to_reward")
+	pickup_coin.picked_up.connect(func(body: Node3D):
+		if (body is Player and body.id == multiplayer.get_unique_id()):
+			QuestManager.update_description("You found the coin! Now go and kill the Target Dummy.")
+			print("picked up coin")
+			state_chart.set_expression_property("item_picked_up", true)
+			state_chart.send_event("to_reward")
 	)
 
 func completed_objectives() -> void:
@@ -55,8 +58,8 @@ func completed_objectives() -> void:
 
 func finished_quest(body: Node3D) -> void:
 	if (body is Player and body.id == multiplayer.get_unique_id()):
-			state_chart.send_event("reward_collected")
-			pickup_zone.body_entered.disconnect(finished_quest)
+		state_chart.send_event("reward_collected")
+		pickup_zone.body_entered.disconnect(finished_quest)
 
 func reward_collected() -> void:
 	QuestManager.hide_quest()
